Added `delete_many` methods for Comments and Logs
The methods are implemented specifically for each backend in their
`Collection`s.

Added methods:
- `delete_all`: Deletes all entities.
- `delete_many`: Delete one or more entities based on a QueryBuilder
  filter.

The intended logic for `delete_many` is:
- When an empty dict is provided, raise ValidationError.
- If the filter provided yields no results, nothing happens.
- Return a list of (former) PKs for the deleted entities.

+++ aiida/orm/implementation/django/logs.py
@@ -8,11 +8,13 @@
 # For further information please visit http://www.aiida.net               #
 ###########################################################################
 """The Django log and log collection module"""
+# pylint: disable=import-error,no-name-in-module
 from __future__ import division
 from __future__ import print_function
 from __future__ import absolute_import
 
-# pylint: disable=import-error,no-name-in-module,fixme
+from django.core.exceptions import ObjectDoesNotExist
+
 from aiida.backends.djsite.db import models
 from aiida.common import exceptions
 
@@ -39,9 +41,9 @@ class DjangoLog(entities.DjangoModelEntity[models.DbLog], BackendLog):
     @property
     def uuid(self):
         """
-        Get the uuid of the object that created the log entry
+        Get the string representation of the uuid of the object that created the log entry
         """
-        return self._dbmodel.uuid
+        return str(self._dbmodel.uuid)
 
     @property
     def time(self):
@@ -95,24 +97,59 @@ class DjangoLogCollection(BackendLogCollection):
         """
         Remove a Log entry from the collection with the given id
 
-        :param log_id: id of the log to delete
+        :param log_id: id of the Log to delete
+        :type log_id: int
+
+        :raises TypeError: if ``log_id`` is not an `int`
+        :raises `~aiida.common.exceptions.NotExistent`: if Log with ID ``log_id`` is not found
         """
-        # pylint: disable=no-name-in-module,import-error
-        from django.core.exceptions import ObjectDoesNotExist
-        assert log_id is not None
+        if not isinstance(log_id, int):
+            raise TypeError("log_id must be an int")
+
         try:
             models.DbLog.objects.get(id=log_id).delete()
         except ObjectDoesNotExist:
             raise exceptions.NotExistent("Log with id '{}' not found".format(log_id))
 
+    def delete_all(self):
+        """
+        Delete all Log entries.
+
+        :raises `~aiida.common.exceptions.IntegrityError`: if all Logs could not be deleted
+        """
+        from django.db import transaction
+        try:
+            with transaction.atomic():
+                models.DbLog.objects.all().delete()
+        except Exception as exc:
+            raise exceptions.IntegrityError("Could not delete all Logs. Full exception: {}".format(exc))
+
     def delete_many(self, filters):
         """
-        Delete all log entries.
-        Delete all log entries in the table if 'filters' is not defined.
-        :param filters: Dictionary, where the keys can be: 'id' or 'node_id'.
-        and the values may be a list. 'time' is not yet implemented.
+        Delete Logs based on ``filters``
+
+        :param filters: similar to QueryBuilder filter
+        :type filters: dict
+
+        :return: (former) ``PK`` s of deleted Logs
+        :rtype: list
+
+        :raises TypeError: if ``filters`` is not a `dict`
+        :raises `~aiida.common.exceptions.ValidationError`: if ``filters`` is empty
         """
+        from aiida.orm import Log, QueryBuilder
+
+        # Checks
+        if not isinstance(filters, dict):
+            raise TypeError("filters must be a dictionary")
         if not filters:
-            models.DbLog.objects.all().delete()
-        else:
-            raise NotImplementedError('Only deleting all by passing an empty filter dictionary is currently supported')
+            raise exceptions.ValidationError("filters must not be empty")
+
+        # Apply filter and delete found entities
+        builder = QueryBuilder().append(Log, filters=filters, project='id').all()
+        entities_to_delete = [_[0] for _ in builder]
+        for entity in entities_to_delete:
+            self.delete(entity)
+
+        # Return list of deleted entities' (former) PKs for checking
+        return entities_to_delete
